/admin - fix: aggressively delete cookies in cs-registration EXEs
Short version: Facebook broke their logout function, then changed its
behavior so it will not log users out if they were not logged in by the
same app trying to log them out. However, this is buggy and sometimes FB
will lose track of the status of the user, and they are stuck logged in.

Previous attempts to nuke all cookies from orbit failed because nw.js
needs special handling for them. In our current nw.js version, many of
the built-in cookie functions are broken and cause the application to
crash.

This commit is a stable workaround this issue. It logs out cookies as it
attempts to delete them, skipping the laravel session cookie.